role management

+++ src/main/java/com/inops/visitorpass/service/impl/EntitlementServiceImpl.java
@@ -5,45 +5,51 @@ import java.util.Optional;
 
 import org.springframework.stereotype.Service;
 
-import com.inops.visitorpass.entity.Entitlement;
+import com.inops.visitorpass.entity.RoleEntitlement;
 import com.inops.visitorpass.repository.EntitlementRepository;
 import com.inops.visitorpass.service.IEntitlement;
 
-@Service("entitlementServices")
+import lombok.RequiredArgsConstructor;
+
+@Service("entitlementService")
+@RequiredArgsConstructor
 public class EntitlementServiceImpl implements IEntitlement {
 
-	private EntitlementRepository entitlementRepository;
-
-	public EntitlementServiceImpl(EntitlementRepository entitlementRepository) {
-		super();
-		this.entitlementRepository = entitlementRepository;
-	}
+	private final EntitlementRepository entitlementRepository;
 
 	@Override
-	public Optional<List<Entitlement>> findAll() {
+	public Optional<List<RoleEntitlement>> findAll() {
 		return Optional.of(entitlementRepository.findAll());
 	}
 
 	@Override
-	public Optional<Entitlement> findById(long id) {
+	public Optional<RoleEntitlement> findById(long id) {
 		return entitlementRepository.findById(id);
 	}
 
 	@Override
-	public Optional<Entitlement> create(Entitlement entitlement) {
+	public Optional<RoleEntitlement> create(RoleEntitlement entitlement) {
 		return Optional.of(entitlementRepository.save(entitlement));
 	}
 
 	@Override
-	public Optional<Entitlement> update(Entitlement entitlement) {
+	public Optional<RoleEntitlement> update(RoleEntitlement entitlement) {
 
 		return Optional.of(entitlementRepository.save(entitlement));
 	}
 
 	@Override
 	public void delete(long entitlementId) {
-		Optional<Entitlement> entitlement = entitlementRepository.findById(entitlementId);
-		entitlementRepository.delete(entitlement.get());
+		Optional<RoleEntitlement> entitlement = entitlementRepository.findById(entitlementId);
+		if (entitlement != null) {
+			entitlementRepository.delete(entitlement.get());
+		}
+	}
+
+	@Override
+	public void deleteAll(List<Long> entitelmentIds) {
+		List<RoleEntitlement> entitlements = entitlementRepository.findAllById(entitelmentIds);
+		entitlementRepository.deleteAll(entitlements);
 	}
 
 }
